spiderlearn-pictures1/requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
Pillow==6.1.0
pyaria2==0.1.1
reportlab==3.5.23
requests==2.22.0
soupsieve==1.9.3
urllib3==1.25.3
